[feat/#160] 프로필 이미지 변경 화면 UI 테스트 작성 및 리팩토링
