+++ lib/cubits/profile/profile_info_cubit.dart
@@ -3,15 +3,12 @@ import 'package:ac_project_app/models/profile/profile.dart';
 import 'package:ac_project_app/provider/api/user/profile_api.dart';
 import 'package:ac_project_app/provider/api/user/user_api.dart';
 import 'package:ac_project_app/util/logger.dart';
-import 'package:ac_project_app/util/string_utils.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class GetProfileInfoCubit extends Cubit<ProfileState> {
   GetProfileInfoCubit() : super(ProfileInitialState()) {
     loadProfileData();
   }
-
-  String? imageNumber;
 
   final userApi = UserApi();
   final profileApi = ProfileApi();
@@ -23,14 +20,13 @@ class GetProfileInfoCubit extends Cubit<ProfileState> {
     final result = await userApi.getUsers();
     result.when(
       success: (user) {
-        imageNumber = user.profileImg;
         emit(
           ProfileLoadedState(
             Profile(
               id: user.id,
               nickname: user.nickname,
               jobGroup: user.jobGroup,
-              profileImage: makeImagePath(user.profileImg),
+              profileImage: user.profileImg,
             ),
           ),
         );
@@ -41,27 +37,11 @@ class GetProfileInfoCubit extends Cubit<ProfileState> {
     );
   }
 
-  void selectImage(int index) {
-    final profile = (state as ProfileLoadedState).profile;
-    emit(ProfileLoadingState());
-
-    imageNumber = '0${index + 1}';
-    emit(
-      ProfileLoadedState(
-        Profile(
-          nickname: profile.nickname,
-          profileImage: makeImagePath(imageNumber!),
-          jobGroup: profile.jobGroup,
-        ),
-      ),
-    );
-  }
-
-  Future<bool> updateProfileImage() async {
-    final result = await profileApi.changeImage(profileImg: imageNumber);
+  Future<bool> updateProfileImage(String profileImage) async {
+    final result = await profileApi.changeImage(profileImg: profileImage);
     return result.when(
       success: (data) {
-        Log.i(data);
+        Log.i(data.toJson());
         return true;
       },
       error: (msg) {
