+++ lib/util/string_utils.dart
@@ -20,9 +20,6 @@ String getSafeTitleText(String? text) {
   return text;
 }
 
-String makeImagePath(String image) =>
-    'assets/images/profile/img_${image}_on.png';
-
 bool isLinkVerified(Link link) =>
     link.image != null &&
     link.image!.isNotEmpty &&
